Password Protected: Fix first-load-iframe-hamster bug.
There is some race condition with calling replaceWith that could cause
the dummy video (the infamous Dramatic Hamster) to show instead of the
real video. This would only happen on first load of the page, so could
only be reproduced in a new Incognito window. Queueing the replacement
in `video.ready` appears to fix it, which is weird because the video
must already be in a ready state, and therefore executes it immediately.